Add files via upload
Build a KNN model using a value of k equals five, find the accuracy on the validation data 
Done

Determine the accuracy for the first 15 values of k the on the validation data:. (1 mark )
66.66%

Determine the minimum value for the parameter that improves results on validation data. 
10 nearest neighbor which increases accuracy from 66.66% to 80%

Building model using Support Vector Machine. Done

Train a logistic regression model and determine the accuracy of the validation data (set C=0.01) 
53.33%

Calculate the F1 score and Jaccard Similarity score for each model from above. Use the Hyperparameter that performed best on the validation data 
Logistic Regression And Support Vector Machine performed much better than other algorithms with accuracy 71% and 80% respectively .